chore: add rstest, pretty_assertions, and nextest config for testing strategy
Add workspace-level dev-dependencies for rstest 0.24 and pretty_assertions 1,
and create .config/nextest.toml for cargo-nextest. Per-crate dev-dependencies
added to weft_reactor, weft_llm, and weft crates for use in later phases of
the testing strategy spec.

Also fixes pre-existing clippy warnings that would have blocked
cargo clippy --workspace --all-targets -- -D warnings.

+++ crates/weft/src/server.rs
@@ -787,7 +787,7 @@ mod tests {
             }],
             budget: BudgetConfig {
                 max_generation_calls: 20,
-                max_iterations: config.gateway.max_command_iterations as u32,
+                max_iterations: config.gateway.max_command_iterations,
                 max_depth: 5,
                 timeout_secs: config.gateway.request_timeout_secs,
                 generation_timeout_secs: config.gateway.request_timeout_secs,
